Refactored HomeScreen class to separate component
