+++ frontend/components/HomeScreen.js
@@ -1,0 +1,100 @@
+import React from 'react';
+import {
+  ActivityIndicator,
+  Button,
+  Clipboard,
+  Image,
+  Share,
+  StatusBar,
+  StyleSheet,
+  Text,
+  TouchableOpacity,
+  View,
+  WebView,
+} from 'react-native';
+import Exponent, { Constants, ImagePicker, registerRootComponent, LinearGradient } from 'expo';
+
+export default class HomeScreen extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = { screen: this.props.screen };
+  }
+
+  _takePhoto = async () => {
+    let pickerResult = await ImagePicker.launchCameraAsync({
+      allowsEditing: false,
+      aspect: [4, 3],
+    });
+    this.props.setScreen('ANALYZE');
+    this._handleImagePicked(pickerResult);
+  };
+
+  _handleImagePicked = async pickerResult => {
+    let uploadResponse, uploadResult;
+    try {
+      this.props.setUploading(true);
+      if (!pickerResult.cancelled) {
+        console.log(1);
+        uploadResponse = await this.uploadImageAsync(pickerResult.uri);
+        console.log(uploadResponse);
+        uploadResult = await uploadResponse.json();
+        this.props.setImage(uploadResult.location);
+      }
+    } catch (e) {
+      console.log({ uploadResponse });
+      console.log({ uploadResult });
+      console.log({ e });
+      alert('Upload failed, sorry :(');
+    } finally {
+      // this.setState({ uploading: false });
+      this.props.setUploading(false);
+    }
+  };
+
+  async uploadImageAsync(uri) {
+    let apiUrl = 'https://file-upload-example-backend-dkhqoilqqn.now.sh/upload';
+
+    let uriParts = uri.split('.');
+    let fileType = uriParts[uriParts.length - 1];
+
+    let formData = new FormData();
+    formData.append('photo', {
+      uri,
+      name: `photo.${fileType}`,
+      type: `image/${fileType}`,
+    });
+
+    let options = {
+      method: 'POST',
+      body: formData,
+      headers: {
+        Accept: 'application/json',
+        'Content-Type': 'multipart/form-data',
+      },
+    };
+
+    return fetch(apiUrl, options);
+  }
+
+  render() {
+    return (
+      <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+        <View style={styles.container}>
+          <LinearGradient colors={['#5161B9', '#9C69CC']} style={{ position: 'absolute', height: 900, width: 400 }} />
+          <TouchableOpacity onPress={this._takePhoto}>
+            <Text style={{ fontSize: 20, color: 'white' }}>TAP TO BEGIN</Text>
+            <Image style={{ width: 150, height: 150 }} source={{ uri: 'https://78.media.tumblr.com/48a0d13c52b402e976bc5d4416552671/tumblr_onew3c4x8a1vxu8n6o1_500.gif' }} />
+          </TouchableOpacity>
+        </View>
+      </View >
+    );
+  }
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
