+++ frontend/package.json
@@ -8,9 +8,7 @@
   "dependencies": {
     "expo": "^20.0.0",
     "react": "16.0.0-alpha.12",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-20.0.0.tar.gz"
-  },
-  "devDependencies": {
+    "react-native": "https://github.com/expo/react-native/archive/sdk-20.0.0.tar.gz",
     "react-devtools": "^3.1.0"
   }
 }
